[CHORE]: Refactorización de archivos de css
