+++ frontend/src/components/FolderTabs.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import "./folderTabs.css";
+import "./css/folderTabs.css";
 
 // {'file1': <div>Contenido 1</div>, 'file2': <div>Contenido 2</div>, 'file3': <div>Contenido 3</div>}
 export default function FolderTabs({sections={}, style}) {
