+++ frontend/src/components/css/delayedButton.css
@@ -1,4 +1,3 @@
-/* App.css */
 .button {
     padding: 10px 20px;
     font-size: 16px;
@@ -32,27 +31,3 @@
       transform: rotate(360deg);
     }
   }
-  
-  .medal {
-    font-size: 2rem;
-}
-
-.medal.gold {
-    color: #FFD700;
-}
-
-.medal.silver {
-    color: #C0C0C0;
-}
-
-.medal.bronze {
-    color: #CD7F32;
-}
-
-.medal.dotted {
-    color: transparent;
-    border: 2px dashed #999;
-    border-radius: 50%;
-    padding: 4px;
-}
-
