+++ frontend/src/components/css/folderTabs.css
@@ -49,7 +49,6 @@
   color: #555;
   font-size: 16px;
   line-height: 1.6;
-  overflow-y: auto;
   height: 90%;
 }
